requirements.txt
Flask[async]==3.1.0
openai==1.57.4
python-dotenv==1.0.1
httpx
